+++ spec/integration/in_browser_specs_spec.rb
@@ -1,21 +1,24 @@
 require 'spec_helper'
 
-feature 'In-browser specs runner' do
-  scenario 'runs all specs', :js do
+describe 'In-browser specs runner', js: true, type: :feature do
+  before do
     visit '/opal_spec'
-    page.should have_content('example_spec ')
-    page.should have_content('requires_opal_spec ')
-    page.should have_content('subdirectory/other_spec ')
-    page.should have_content('3 examples, 0 failures')
   end
 
-  scenario "runs single spec file", :js do
-    visit '/opal_spec'
-    click_link 'subdirectory/other_spec'
+  it 'all specs' do
+    expect(page).to have_content('3 examples, 0 failures')
+    expect(page).to have_content('example_spec ')
+    expect(page).to have_content('requires_opal_spec ')
+    expect(page).to have_content('subdirectory/other_spec ')
+  end
 
-    page.should_not have_content('example_spec ')
-    page.should_not have_content('requires_opal_spec ')
-    page.should have_content('subdirectory/other_spec ')
-    page.should have_content('1 examples, 0 failures')
+  it 'single spec file' do
+    click_link 'subdirectory/other_spec'
+    expect(page).to have_content('1 example, 0 failures')
+    expect(page).to_not have_content('example_spec ')
+    expect(page).to_not have_content('requires_opal_spec ')
+    expect(page).to have_content('subdirectory/other_spec ')
+    expect(page).to have_link 'All specs'
   end
 end
+